Move pick experiment quick pick
