+++ extension/src/experiments/quickPick.ts
@@ -1,42 +1,6 @@
-import { Experiment } from './webview/contract'
-import { getColumnPathsQuickPickDetail } from './model/util'
 import { GcPreserveFlag, QueueRemoveFlag } from '../cli/dvc/constants'
-import { quickPickManyValues, quickPickValue } from '../vscode/quickPick'
+import { quickPickManyValues } from '../vscode/quickPick'
 import { Title } from '../vscode/title'
-import { Toast } from '../vscode/toast'
-
-export type ExperimentWithName = Experiment & {
-  name?: string
-}
-
-export const pickExperiment = (
-  experiments: ExperimentWithName[],
-  firstThreeColumnOrder: string[],
-  title: Title = Title.SELECT_EXPERIMENT
-): Thenable<{ id: string; name: string } | undefined> | undefined => {
-  if (experiments.length === 0) {
-    void Toast.showError('There are no experiments to select.')
-  } else {
-    return quickPickValue<{ id: string; name: string }>(
-      experiments.map(experiment => {
-        const { label, id, name, displayNameOrParent } = experiment
-        return {
-          description: displayNameOrParent,
-          detail: getColumnPathsQuickPickDetail(
-            experiment,
-            firstThreeColumnOrder
-          ),
-          label,
-          value: {
-            id,
-            name: name || label
-          }
-        }
-      }),
-      { matchOnDescription: true, matchOnDetail: true, title }
-    )
-  }
-}
 
 export const pickGarbageCollectionFlags = () =>
   quickPickManyValues<GcPreserveFlag>(
